Add files via upload
Dear Dr

Thank you for your invaluable feedback today. Areej shared with us the notes, and we promptly took action to address the suggestions provided.

Regarding the expansion of the Kernel Density Estimation Heatmap of Merged Data, we have incorporated a color bar and added explanations following the graphs to enhance clarity.

Concerning the Heatmap of Aggregated Air Quality Metrics, we eagerly await the file to further our project post-finals.

We have plotted taxi_density versus PM2.5 in the Spaghetti Plot.

In the Spatial Autocorrelation analysis, we have included additional code to study the effect of temperature.

With these adjustments, our coding journey for this project concludes. While it was indeed a lengthy process, your guidance, and teamwork were pivotal in our achievements. Thank you once again for your support. We are committed to continuing our work on the Journal and submitting it before the deadline.